requirements.txt
Django~=4.1.7
python-dotenv~=0.21.1
django-filter~=22.1
djangorestframework~=3.14.0
psycopg2 ~=2.9.5